+++ src/pages/Main.tsx
@@ -1,15 +1,21 @@
 import { Flex } from '@chakra-ui/react';
 
-import Filter from '../components/Filter';
+import Cart from '../components/Cart';
 import Header from '../components/Header';
 import Product from '../components/Product';
+import SpaceFilter from '../components/SpaceFilter';
 
 function Main() {
   return (
-    <Flex flexDirection={'column'} margin={10}>
+    <Flex flexDirection={'column'} m={10}>
       <Header />
-      <Filter />
-      <Product />
+      <Flex flexDirection={'column'} mx={10}>
+        <Flex my={10} justifyContent={'space-between'}>
+          <SpaceFilter />
+          <Cart />
+        </Flex>
+        <Product />
+      </Flex>
     </Flex>
   );
 }
